suggested changes

+++ lib/service_monitor_utils.cc
@@ -264,7 +264,7 @@ double calculate_cpu_usage(unsigned long long oldCpuTime, unsigned long long new
       (newProcessTime.uTime - oldProcessTime.uTime) + (newProcessTime.sTime - oldProcessTime.sTime);
   unsigned long long cpuTimeDelta = newCpuTime - oldCpuTime;
 
-  double cpuUsage = 100.0 * (double)processTimeDelta / (double)cpuTimeDelta * numCores;
+  double cpuUsage = (100.0 * processTimeDelta / cpuTimeDelta) * numCores;
 
   return cpuUsage;
 }
